Implemented a basic macros for logging module:
IHFT_LOG_DEBUG(...)
IHFT_LOG_INFO(...)
IHFT_LOG_WARN(...)
IHFT_LOG_ERROR(...)

Example:

    IHFT_LOG_INFO("Hello {} !!!", "world");

->

UTC 2023-06-14 16:28:26.457495 INFO test_logger.cpp(13):void C_A_T_C_H_T_E_S_T_0() Hello world !!!

+++ platform/include/platform/platform.h
@@ -32,7 +32,7 @@ struct trait
     static unsigned get_total_cpus() noexcept;
 
     // Get current thread native id
-    static long long get_thread_id() noexcept;
+    static long get_thread_id() noexcept;
 
     // This call locks all pages mapped into the address space of the calling process.
     // This includes the pages of the code, data, and stack segment,
